refactor: 테스트 관련 유틸 함수 폴더 구조 변경
